+++ src/main/resources/db/migration/V3__create_user_roles.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS user_roles (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
 
     user_id BIGINT NOT NULL,
     role_id BIGINT NOT NULL,
@@ -8,6 +8,6 @@ CREATE TABLE IF NOT EXISTS user_roles (
     updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     deleted_at TIMESTAMP,
 
-    CONSTRAINT fk_user FOREIGN KEY (user_id) REFERENCES users(id),
-    CONSTRAINT fk_role FOREIGN KEY (role_id) REFERENCES roles(id)
+    CONSTRAINT fk_user FOREIGN KEY (user_id) REFERENCES users (id),
+    CONSTRAINT fk_role FOREIGN KEY (role_id) REFERENCES roles (id)
 );
